+++ wiki/public/manifest.json
