+++ wiki/src/index.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { BrowserRouter as Router} from 'react-router-dom'
+
+import App from './components/App';
+
+import 'bootstrap/dist/css/bootstrap.css';
+import './index.css';
+
+ReactDOM.render(
+  <React.StrictMode>
+  <Router>
+    <App />
+  </Router>  
+  </React.StrictMode>,
+  document.getElementById('root')
+);
